[BE] Feat: NPoint 기능 개발

See merge request s11-webmobile2-sub2/S11P12D207!22

+++ backend/ndolphin/src/main/java/com/web/ndolphin/domain/User.java
@@ -1,14 +1,22 @@
 package com.web.ndolphin.domain;
 
-import jakarta.persistence.*;
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.OneToOne;
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
-
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
 @Getter
@@ -30,7 +38,7 @@ public class User {
 
     private String mbti;
 
-    private int nPoint;
+    private Long nPoint;
 
     @Enumerated(EnumType.STRING)
     private LoginType type;
